parseInt to parseFloat and comments

+++ population_density.js
@@ -10,17 +10,18 @@ function parseCSV(path) {
 
     for (var i = 0; i < array.length; i++) {
         for (var j = 0; j < array[i].length; j++) {
-            array[i][j] = parseInt(array[i][j]);
+            array[i][j] = parseFloat(array[i][j]);
         }
     }
     return array;
 }
 
 function getPopulation(density, lat, long) {
+    // lat.long to pixel
     var pixel_lat = 900 - (lat * 10);
     var pixel_long = 1800 + (long * 10);
     pixel_lat = parseInt(pixel_lat);
     pixel_long = parseInt(pixel_long);
 
-    return density[pixel_lat][pixel_long]; // persons/km^2
+    return density[pixel_lat][pixel_long]; // unit is persons/km^2
 }
